+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#848 - Add ready-to-use config files to the Milvus repo for enhanced user experince
 - \#860 - Remove redundant checks in CacheMgr's constructor
 - \#908 - Move "primary_path" and "secondary_path" to storage config
+- \#931 - Remove "collector" from config
 
 ## Task